+++ src/onegov.core/onegov/core/orm/mixins/__init__.py
@@ -2,6 +2,7 @@ from onegov.core.orm.mixins.content import (
     content_property,
     ContentMixin,
     data_property,
+    dict_property,
     meta_property,
 )
 from onegov.core.orm.mixins.timestamp import TimestampMixin
@@ -10,6 +11,7 @@ __all__ = [
     'content_property',
     'ContentMixin',
     'data_property',
+    'dict_property',
     'meta_property',
-    'TimestampMixin'
+    'TimestampMixin',
 ]